backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pymongo==4.7.3
python-nmap==0.7.1
Werkzeug==3.1.3
python-openvpn==0.8.8
